started a main for the C side

+++ operations.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-double divide(double num, double den) {
+int divide(int num, int den) {
     if (den == 0) {
         printf("Error! Division by zero is not allowed.\n");
         return -1;
@@ -8,16 +8,16 @@ double divide(double num, double den) {
     return num / den;
 }
 
-double add(double a, double b) {
+int add(int a, int b) {
     return a + b;
 }
 
-double subtract(double a, double b) {
+int subtract(int a, int b) {
     return a - b;
 }
 
-double multiply(double a, double b) {
-    double result = 0;
+int multiply(int a, int b) {
+    int result = 0;
     if (b < 0) b = b * -1;
     for (int i = 0; i < b; i++) {
         result = add(result, a);
